% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example data.frame}
\format{
A data frame with 2 columns:
\describe{
\item{x}{some numbers}
\item{y}{some other numbers}
}
}
\usage{
dat
}
\description{
Example data.frame
}
\keyword{datasets}
